@@ -5,10 +5,10 @@
     "keywords": ["maxmind geoip"],
     "license": "BSD-2-Clause",
     "require": {
-        "php": ">=5.3 || ^7.0",
-        "symfony/framework-bundle": "~2.3 || ~3.0 || ^4.0",
-        "symfony/dependency-injection": "~2.3 || ~3.0 || ^4.0",
-        "symfony/console": "~2.3 || ~3.0 || ^4.0"
+        "php": "^7.0",
+        "symfony/framework-bundle": "^3.4 || ^4.0",
+        "symfony/dependency-injection": "^3.4 || ^4.0",
+        "symfony/console": "^3.4 || ^4.0"
     },
     "require-dev": {
         "phpunit/phpunit": "~5.7"
